+++ _hijack_root/qml/Map/MapEditorBaseView.qml
@@ -94,23 +94,21 @@ ScrollView {
                 height: body.height / scale
                 scale: body.tileScale
                 visible: true // Hardcode enable showing parallax, since you can just change parallax to none to display none.
-                z: -2
+                z: -2 // Reduce to -2 so it appears before Tiled Maps
             }
 
             // Grab map images from the output folder of OMORI Map Renderer.
             Image {
                 property string imageName: "map" + DataManager.makeIdText(root.mapId, 2)
-                property string folder: DataManager.projectUrl + "scaled"
+                property string folder: DataManager.projectUrl + "render"
                 source: root.mapId > 0 ? folder + "/" + encodeURIComponent(imageName) + ".png" : ""
-                fillMode: Image.Tile
-                horizontalAlignment: Image.AlignLeft
-                verticalAlignment: Image.AlignTop
+                fillMode: Image.Stretch // Changed to Stretch to help prevent crashing with very large maps
                 x: (body.width - width) / 2
                 y: (body.height - height) / 2
                 width: body.width / scale
                 height: body.height / scale
                 scale: body.tileScale
-                visible: root.parallaxShow // Change Show In Editor option to apply to tiled maps.
+                visible: root.parallaxShow // Change Show In Editor option to apply to Tiled Maps.
                 z: -1
             }